refactor: remove duplicate sorting in various places
